requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg[binary]
python-dotenv
pandas

pytest
httpx
